requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
cftime==1.6.4.post1
click==8.2.1
clingo==5.8.0
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dask==2025.7.0
debugpy==1.8.16
decorator==5.2.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d38d7161247e64276d4c44d9d0605291e80bd969
executing==2.2.0
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
fvcore==0.1.5.post20221221
grpcio==1.74.0
hydra-core==1.3.2
imageio==2.37.0
iniconfig==2.1.0
iopath==0.1.9
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
locket==1.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
popper-ilp @ https://github.com/logic-and-learning-lab/Popper/archive/refs/tags/v3.1.0.tar.gz#sha256=c9504c120f3edde07881943eae5bff06f5bc386f2423abf8465e425f9e3d3ffc
portalocker==3.2.0
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.10
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pysat==3.2.2
pyswip==0.3.3
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
ruff==0.12.9
scikit-image==0.25.2
scipy==1.16.1
setuptools<81   # important!
six==1.17.0
split-folders==0.5.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.6.11
toolz==1.0.0
torch==2.8.0
torch-tb-profiler==0.4.3
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
xarray==2025.8.0
xmltodict==0.14.2
yacs==0.1.8
z3-solver==4.15.3.0
